Refactor & Styles: Fix sidebar hover animation
The sidebar hover animation was looking a bit buggy,
because the ::before pseudo-element on the sidebar-control
was appearing a little bit before the other elements when
hovered, which caused a "glitchy effect".

I set all the transitions to 0ms for the whole row
to recolor when hovered immediately, which gives a
cleaner effect.